requirements.txt
simplejson==3.10.0
trello==0.9.1
flake8==3.2.1
pep8==1.7.0
phabricator==0.7.0
PyYAML==3.12
